+++ src/models/ComponentStyle.js
@@ -3,7 +3,10 @@ import hashStr from '../vendor/glamor/hash'
 
 import type { RuleSet, NameGenerator, Flattener, Stringifier } from '../types'
 import StyleSheet from './StyleSheet'
+import { IS_BROWSER } from '../constants'
 import isStyledComponent from '../utils/isStyledComponent'
+
+const areStylesCacheable = IS_BROWSER
 
 const isStaticRules = (rules: RuleSet, attrs?: Object): boolean => {
   for (let i = 0; i < rules.length; i += 1) {
@@ -46,6 +49,9 @@ export default (
   flatten: Flattener,
   stringifyRules: Stringifier
 ) => {
+  /* combines hashStr (murmurhash) and nameGenerator for convenience */
+  const generateRuleHash = (str: string) => nameGenerator(hashStr(str))
+
   class ComponentStyle {
     rules: RuleSet
     componentId: string
@@ -56,10 +62,12 @@ export default (
       this.rules = rules
       this.isStatic = !isHRMEnabled && isStaticRules(rules, attrs)
       this.componentId = componentId
-      if (!StyleSheet.instance.hasInjectedComponent(this.componentId)) {
+
+      if (!StyleSheet.master.hasId(componentId)) {
         const placeholder =
-          process.env.NODE_ENV !== 'production' ? `.${componentId} {}` : ''
-        StyleSheet.instance.deferredInject(componentId, true, [placeholder])
+          process.env.NODE_ENV !== 'production' ? [`.${componentId} {}`] : []
+
+        StyleSheet.master.deferredInject(componentId, placeholder)
       }
     }
 
@@ -69,42 +77,25 @@ export default (
      * Returns the hash to be injected on render()
      * */
     generateAndInjectStyles(executionContext: Object, styleSheet: StyleSheet) {
-      const { isStatic, lastClassName } = this
-      if (isStatic && lastClassName !== undefined) {
+      const { isStatic, componentId, lastClassName } = this
+      if (areStylesCacheable && isStatic && lastClassName !== undefined) {
         return lastClassName
       }
 
       const flatCSS = flatten(this.rules, executionContext)
-      const hash = hashStr(this.componentId + flatCSS.join(''))
+      const name = generateRuleHash(this.componentId + flatCSS.join(''))
 
-      const { stylesCacheable } = styleSheet
-      const existingName = styleSheet.getName(hash)
-
-      if (existingName !== undefined) {
-        if (stylesCacheable) {
-          this.lastClassName = existingName
-        }
-        return existingName
+      if (!styleSheet.hasNameForId(componentId, name)) {
+        const css = stringifyRules(flatCSS, `.${name}`)
+        styleSheet.inject(this.componentId, css, name)
       }
 
-      const name = nameGenerator(hash)
-      if (stylesCacheable) {
-        this.lastClassName = existingName
-      }
-      if (styleSheet.alreadyInjected(hash, name)) {
-        return name
-      }
-
-      const css = stringifyRules(flatCSS, `.${name}`)
-      // NOTE: this can only be set when we inject the class-name.
-      // For some reason, presumably due to how css is stringifyRules behaves in
-      // differently between client and server, styles break.
-      styleSheet.inject(this.componentId, true, css, hash, name)
+      this.lastClassName = name
       return name
     }
 
-    static generateName(str: string) {
-      return nameGenerator(hashStr(str))
+    static generateName(str: string): string {
+      return generateRuleHash(str)
     }
   }